major cart update

+++ src/modules/restaurant-cart/restaurant-cart.service.ts
@@ -42,19 +42,26 @@ export class RestaurantCartService {
   async addMealToCart(req: AddMealRestaurantCartRequest) {
     return await this.addMealTransaction.run(req);
   }
+
   async getCartMeals() {
     const cart = await this.restaurantCartRepository.findOne({
       where: { user_id: this.request.user.id },
       relations: {
         restaurant_cart_meals: {
-          meal: {offer:true},
-          cart_meal_options: { meal_option_price:{option: {option_group:true} }},
+          meal: { offer: true },
+          cart_meal_options: {
+            meal_option_price: {
+              meal_option_group: true,
+              option: { option_group: true },
+            },
+          },
         },
         restaurant: true,
       },
     });
 
-    // delete cart meals if not active
+    if (!cart) return null;
+
     for (let i = cart.restaurant_cart_meals?.length - 1; i >= 0; i--) {
       if (!cart.restaurant_cart_meals[i]?.meal?.is_active) {
         await this.restaurantCartMealRepository.delete({
@@ -62,10 +69,11 @@ export class RestaurantCartService {
         });
       }
     }
-     
+
     const default_address = await this.addressRepository.findOne({
       where: { user_id: this.request.user.id, is_favorite: true },
     });
+
     const settings = await this.constantRepository.find();
     const fixed_delivery_fee = settings.find(
       (s) => s.type === ConstantType.FIXED_DELIVERY_FEE,
@@ -76,10 +84,12 @@ export class RestaurantCartService {
     const fixed_delivery_distance = settings.find(
       (s) => s.type === ConstantType.FREE_DELIVERY_DISTANCE,
     ).variable;
+
     const distance = calculateDistances(
       [cart.restaurant.latitude, cart.restaurant.longitude],
       [default_address.latitude, default_address.longitude],
     );
+
     const delivery_fee =
       Number(fixed_delivery_fee) +
       (Number(distance) - Number(fixed_delivery_distance) < 0
@@ -87,72 +97,84 @@ export class RestaurantCartService {
         : Number(distance) - Number(fixed_delivery_distance)) *
         Number(delivery_price_per_km);
 
-    if (!cart) return null;
-    
     const response = plainToInstance(
       GetCartMealsResponse,
       cart.restaurant_cart_meals.map((m) => {
-        // Check if the meal has a valid offer
         const offer = m.meal.offer;
         let discount_percentage = 0;
         let discounted_price = Number(m.meal.price);
-    
-        if (
+        const isOfferActive =
           offer &&
           offer.is_active &&
-          new Date(offer.start_date) <= new Date(new Date().setUTCHours(new Date().getUTCHours() + 3)) && // Yemen Time (UTC+3)
-          new Date(offer.end_date) > new Date(new Date().setUTCHours(new Date().getUTCHours() + 3))
-        ) {
+          new Date(offer.start_date) <=
+            new Date(new Date().setUTCHours(new Date().getUTCHours() + 3)) &&
+          new Date(offer.end_date) >
+            new Date(new Date().setUTCHours(new Date().getUTCHours() + 3));
+
+        if (isOfferActive) {
           discount_percentage = Number(offer.discount_percentage);
-          discounted_price = discounted_price - (discounted_price * discount_percentage) / 100;
+          discounted_price =
+            discounted_price - (discounted_price * discount_percentage) / 100;
         }
-    
-        // Calculate total unit price with options
+
+        const options = m.cart_meal_options.map((o) => {
+          const original_price = o.meal_option_price.price;
+          let discounted_option_price = original_price;
+
+          const group_apply_offer =
+            o.meal_option_price.meal_option_group?.apply_offer;
+
+          if (isOfferActive && group_apply_offer) {
+            discounted_option_price =
+              original_price - (original_price * discount_percentage) / 100;
+          }
+
+          return {
+            ...o,
+            original_price,
+            discounted_price: discounted_option_price,
+            price: discounted_option_price,
+            option: o.meal_option_price.option,
+
+            option_group: o.meal_option_price.option.option_group,
+          };
+        });
+
         const total_unit_price =
           discounted_price +
-          Number(
-            m.cart_meal_options.reduce(
-              (acc, curr) => acc + curr.meal_option_price.price,
-              0,
-            ),
-          );
-    
+          Number(options.reduce((acc, o) => acc + o.discounted_price, 0));
+
         return {
           ...m.meal,
           meal_id: m.meal.id,
           id: m.id,
           quantity: m.quantity,
-          total_price: total_unit_price , // Multiply by quantity for final total
-          options: m.cart_meal_options.map((o) => ({
-            ...o,
-            price: o.meal_option_price.price,
-            option_group: o.meal_option_price.option.option_group,
-          })),
-         
-        
+          total_price: total_unit_price,
+          options,
         };
       }),
-      {
-        excludeExtraneousValues: true,
-      },
+      { excludeExtraneousValues: true },
     );
-    
-    const restaurant_respone = plainToInstance(
+
+    const restaurant_response = plainToInstance(
       RestaurantResponse,
       cart.restaurant,
       { excludeExtraneousValues: true },
     );
+
     return {
       meals: response,
-      restaurant: restaurant_respone,
+      restaurant: restaurant_response,
       delivery_fee: delivery_fee,
-      wallet:await this.transactionService.getWallet(this.request.user.id)
+      wallet: await this.transactionService.getWallet(this.request.user.id),
     };
   }
+
   async clearCart() {
     const cart = await this.restaurantCartRepository.findOne({
       where: { user_id: this.request.user.id },
     });
+
     const cart_items = await this.restaurantCartMealRepository.find({
       where: { cart_id: cart.id },
     });
@@ -161,26 +183,71 @@ export class RestaurantCartService {
   }
 
   async deleteCartMeal(cart_meal_id: string) {
-    return await this.restaurantCartMealRepository.delete({
-      id: cart_meal_id,
-    });
+    return await this.restaurantCartMealRepository.delete({ id: cart_meal_id });
   }
+
   async updateCartMeal(req: UpdateCartMealRequest) {
     const response = await this.updateMealRestaurantCartTransaction.run(req);
-    const cart_meals = this.getCartMeals();
-    return (await cart_meals).meals.find(
-      (cart_meal) => cart_meal.id === response.id,
-    );
+    const cart_meals = await this.getCartMeals();
+    return cart_meals.meals.find((cart_meal) => cart_meal.id === response.id);
   }
 
   async getCartMealDetails(cart_meal_id: string) {
     const cart_meal = await this.restaurantCartMealRepository.findOne({
       where: { id: cart_meal_id },
       relations: {
-        meal: { meal_option_groups: { option_group: { options: true }  },offer:true },
-        cart_meal_options: { meal_option_price: { option: true } },
+        meal: {
+          meal_option_groups: {
+            meal_option_prices: { option: true },
+            option_group: { options: true },
+          },
+          offer: true,
+        },
+        cart_meal_options: {
+          meal_option_price: { option: true, meal_option_group: true },
+        },
       },
     });
+
+    const offer = cart_meal.meal.offer;
+    const now = new Date(new Date().setUTCHours(new Date().getUTCHours() + 3));
+    const isOfferActive =
+      offer &&
+      offer.is_active &&
+      new Date(offer.start_date) <= now &&
+      new Date(offer.end_date) > now;
+
+    const discount_percentage = isOfferActive
+      ? Number(offer.discount_percentage)
+      : 0;
+    let discounted_meal_price = cart_meal.meal.price;
+
+    if (isOfferActive) {
+      discounted_meal_price -=
+        (discounted_meal_price * discount_percentage) / 100;
+    }
+
+    const options = cart_meal.cart_meal_options.map((opt) => {
+      const original_price = opt.meal_option_price?.price;
+      let final_price = original_price;
+
+      if (
+        isOfferActive &&
+        opt.meal_option_price.meal_option_group?.apply_offer
+      ) {
+        final_price -= (final_price * discount_percentage) / 100;
+      }
+
+      return {
+        ...opt,
+        price: final_price,
+      };
+    });
+
+    const total_price =
+      Number(discounted_meal_price) +
+      Number(options.reduce((acc, opt) => acc + opt.price, 0));
+
     const response = plainToInstance(
       GetCartMealsResponse,
       {
@@ -188,25 +255,23 @@ export class RestaurantCartService {
         meal_id: cart_meal.meal.id,
         id: cart_meal.id,
         quantity: cart_meal.quantity,
-        total_price:
-          Number(cart_meal.meal.price) +
-          Number(
-            cart_meal.cart_meal_options.reduce(
-              (acc, curr) => acc + curr.meal_option_price?.price,
-              0,
-            ),
-          ),
+        total_price: total_price,
       },
       { excludeExtraneousValues: true },
     );
-    //check if option is selected
-    response.option_groups.forEach((option_group) => {
-      option_group.options.forEach((option) => {
+    console.log(
+      cart_meal.cart_meal_options.map((o) => o.meal_option_price.option_id),
+    );
+    response?.option_groups?.forEach((option_group) => {
+      option_group?.options?.forEach((option) => {
+        console.log('option', option.option_id);
         option.is_selected = cart_meal.cart_meal_options.some(
-          (cart_meal_option) => cart_meal_option.meal_option_price.option.id === option.id,
+          (cart_meal_option) =>
+            cart_meal_option.meal_option_price.option.id == option.id,
         );
       });
     });
+
     return response;
   }
 
@@ -216,48 +281,59 @@ export class RestaurantCartService {
       relations: {
         restaurant_cart_meals: {
           meal: { offer: true },
-          cart_meal_options: { meal_option_price: { option: true } },
+          cart_meal_options: {
+            meal_option_price: {
+              meal_option_group: true,
+              option: true,
+            },
+          },
         },
       },
     });
 
-    
+    if (!cart) {
+      return { meals_count: 0, total_price: 0 };
+    }
 
-    
-      if (!cart) {
-        return { meals_count: 0, total_price: 0 };
+    const now = new Date(new Date().setUTCHours(new Date().getUTCHours() + 3));
+
+    const total_price = cart.restaurant_cart_meals.reduce((acc, cartMeal) => {
+      let mealPrice = cartMeal.meal.price;
+      const offer = cartMeal.meal.offer;
+      let discount = 0;
+
+      if (
+        offer &&
+        offer.is_active &&
+        new Date(offer.start_date) <= now &&
+        new Date(offer.end_date) > now
+      ) {
+        discount = Number(offer.discount_percentage);
+        mealPrice = mealPrice - (mealPrice * discount) / 100;
       }
-      const nowUtc = new Date();
-      nowUtc.setHours(nowUtc.getHours() + 3);
-      const nowInYemenTime = nowUtc;
-      
-      const data = {
-        meals_count: cart.restaurant_cart_meals.length,
-        total_price: cart.restaurant_cart_meals.reduce((acc, cartMeal) => {
-          let mealPrice = cartMeal.meal.price;
-      
-          const offer = cartMeal.meal.offer;
-          if (
-            offer &&
-            offer.is_active &&
-            new Date(offer.start_date) <= nowInYemenTime &&
-            new Date(offer.end_date) > nowInYemenTime
-          ) {
-            const discountPercentage = Number(offer.discount_percentage) || 0;
-            mealPrice = mealPrice - (mealPrice * discountPercentage) / 100;
-          }
-      
-          const optionsTotal = cartMeal.cart_meal_options.reduce(
-            (optionsAcc, optionItem) => optionsAcc + optionItem.meal_option_price?.price,
-            0,
-          );
-      
-          return acc + cartMeal.quantity * (mealPrice + optionsTotal);
-        }, 0),
-      };
-      
-      return data;
-      
-    }    
 
+      const optionsTotal = cartMeal.cart_meal_options.reduce((optAcc, opt) => {
+        let optionPrice = opt.meal_option_price.price;
+
+        if (
+          discount > 0 &&
+          opt.meal_option_price.meal_option_group?.apply_offer
+        ) {
+          optionPrice -= (optionPrice * discount) / 100;
+        }
+
+        return optAcc + optionPrice;
+      }, 0);
+
+      return (
+        Number(acc) +
+        cartMeal.quantity * Number(Number(mealPrice) + Number(optionsTotal))
+      );
+    }, 0);
+
+    return {
+      meals_count: cart.restaurant_cart_meals.length,
+      total_price,
+    };
   }
+}
